Split vespa_zkfacade__restrict on space too
